refactoring: Rename Components
