+++ components/Input.js
@@ -1,0 +1,23 @@
+import React, { Component } from 'react';
+import { AppRegistry, TextInput } from 'react-native';
+
+export default class UselessTextInput extends Component {
+  constructor(props) {
+    super(props);
+    this.state = { text: 'PomoPomo' };
+  }
+
+  render() {
+    const { text } = this.state;
+    return (
+      <TextInput
+        style={{ height: 40, borderColor: 'gray', borderWidth: 1 }}
+        onChangeText={text => this.setState({ text })}
+        value={text}
+      />
+    );
+  }
+}
+
+// skip this line if using Create React Native App
+AppRegistry.registerComponent('AwesomeProject', () => UselessTextInput);
