+++ components/PlanBarItem.js
